create module for GridFSConnector
